add abstract job
